+++ src/main/java/com/cmeza/sdgenerator/support/maker/values/CommonValues.java
@@ -1,10 +1,14 @@
 package com.cmeza.sdgenerator.support.maker.values;
 
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
 /**
  * Created by carlos on 08/04/17.
  */
+@AllArgsConstructor
 public enum CommonValues {
-
+    NONE(""),
     SPACE(" "),
     NEWLINE("\r\n"),
     TAB("\t"),
@@ -21,15 +25,8 @@ public enum CommonValues {
     DIAMOND_START("<"),
     DIAMOND_END(">");
 
-    private String value;
-
-    CommonValues(String value) {
-        this.value = value;
-    }
-
-    public String getValue() {
-        return value;
-    }
+    @Getter
+    private final String value;
 
     @Override
     public String toString() {