+++ src/main/java/com/cmeza/sdgenerator/support/maker/values/ScopeValues.java
@@ -1,24 +1,21 @@
 package com.cmeza.sdgenerator.support.maker.values;
 
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
 import static com.cmeza.sdgenerator.support.maker.values.CommonValues.SPACE;
 
 /**
  * Created by carlos on 08/04/17.
  */
+@AllArgsConstructor
 public enum ScopeValues {
 
     PUBLIC("public" + SPACE),
     PRIVATE("private" + SPACE);
 
-    private String value;
-
-    ScopeValues(String value) {
-        this.value = value;
-    }
-
-    public String getValue() {
-        return value;
-    }
+    @Getter
+    private final String value;
 
     @Override
     public String toString() {